requirements.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
pypdf
pandas
streamlit